requirements.txt
altgraph==0.17
appdirs==1.4.4
asgiref==3.3.1
auto-py-to-exe==2.7.5
bcrypt==3.1.7
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
distlib==0.3.0
dj-database-url==0.5.0
Django==3.1.3
django-heroku==0.3.1
django-tinymce4-lite==1.8.0
dlib==19.20.0
docutils==0.16
Eel==0.12.4
filelock==3.0.12
Flask==1.1.2
Flask-Bcrypt==0.7.1
future==0.18.2
gevent==20.6.2
gevent-websocket==0.10.1
gitdb==4.0.5
GitPython==3.1.3
greenlet==0.4.16
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.6.1
imutils==0.5.3
itsdangerous==1.1.0
jedi==0.17.1
Jinja2==2.11.2
joblib==0.17.0
jsmin==2.2.2
MarkupSafe==1.1.1
nest-asyncio==1.3.3
numpy==1.19.2
opencv-contrib-python==4.2.0.34
opencv-python==4.2.0.34
pandas==1.0.5
parso==0.7.0
pathtools==0.1.2
pefile==2019.4.18
pep517==0.6.0
pexe37==0.9.6.5
pexpect==4.8.0
phue==1.1
Pillow==7.2.0
prompt-toolkit==3.0.5
psycopg2==2.8.6
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.6.1
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2020.1
qPython==2.0.0
questionary==1.5.2
requests==2.24.0
rtsp==1.1.8
Rx==3.1.1
scikit-learn==0.23.2
scipy==1.5.4
sh==1.13.1
six==1.15.0
sklearn==0.0
smmap==3.0.4
sqlparse==0.4.1
threadpoolctl==2.1.0
toml==0.10.1
urllib3==1.25.9
virtualenv==20.0.25
watchdog==0.10.3
wcwidth==0.2.4
Werkzeug==1.0.1
whichcraft==0.6.1
whitenoise==5.2.0
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0